+++ CHANGELOG.md
@@ -1,9 +1,9 @@
 # Changelog
 # Version 1.1.0
 - Added Compendiums for SF2e
-- The SF2e deities with avatar definitions get their spell effects
+- The SF2e deities with avatar definitions get their Avatar spell effects
 - If you use the PF2e anachronism module, the PF2e deities effects will work in SF2e.
-- If a deity has both a PF2e and a SF2e version and avatar, the avatar for the current system will be used.
+- If a deity has both a PF2e and a SF2e version, the avatar specific to the current system will be used.
 
 # Version 1.0.1
 - Refactoring and fixing a few issues